Create homework_assignment_2
1. Ask the user to input 2-dimensional (x,y) coordinates and store the input in a list of tuples, i.e.
[ (1.0,2.5), (3,4), (5,6) ]. Note that coordinate values can be decimal values. Hint: Ask how many
numbers will be given first.
2. Calculate the centre of mass for the given set of coordinates, i.e. ( 𝑥̅,𝑦̅ ) and display it.
3. Calculate the distance between each coordinate ant the center of mass using an Euclidean
distance formula, and store the distances in a list. Note: You should learn how to calculate a
square root using built-in Python math functions. You may also like to use the built-in zip
function that processes tuples and lists.
4. Using the list of distances find the coordinates which are closest and farthest away from the
center of mass, and then display them, along with the distance to center of mass. Hint: When
you process the list of distances, if you get an index, the same index may be used in the list of
coordinates as well.